LPS-28940 Leave YouTube specific parts back to YouTubeDisplayContext class
