+++ modules/apps/youtube/youtube-web/src/main/java/com/liferay/youtube/web/internal/display/context/BaseVideoEmbedderDisplayContext.java
@@ -43,38 +43,7 @@ public class BaseVideoEmbedderDisplayContext implements VideoEmbedderDisplayCont
 		StringBundler sb = new StringBundler(13);
 
 		sb.append(HttpUtil.getProtocol(_request));
-		sb.append("://www.youtube.com/embed/");
-		sb.append(getId());
-		sb.append("?wmode=transparent");
-
-		if (isAutoPlay()) {
-			sb.append("&amp;autoplay=1");
-		}
-
-		if (isClosedCaptioning()) {
-			sb.append("&amp;cc_load_policy=1");
-		}
-
-		if (!isEnableKeyboardControls()) {
-			sb.append("&amp;disablekb=1");
-		}
-
-		if (isAnnotations()) {
-			sb.append("&amp;iv_load_policy=1");
-		}
-		else {
-			sb.append("&amp;iv_load_policy=3");
-		}
-
-		if (isLoop()) {
-			sb.append("&amp;loop=1&amp;playlist=");
-			sb.append(getId());
-		}
-
-		if (Validator.isNotNull(getStartTime())) {
-			sb.append("&amp;start=");
-			sb.append(getStartTime());
-		}
+		sb.append("stub for now");
 
 		return sb.toString();
 	}
@@ -111,17 +80,6 @@ public class BaseVideoEmbedderDisplayContext implements VideoEmbedderDisplayCont
 		return _id;
 	}
 
-	public String getImageURL() {
-		StringBundler sb = new StringBundler(4);
-
-		sb.append(HttpUtil.getProtocol(_request));
-		sb.append("://img.youtube.com/vi/");
-		sb.append(getId());
-		sb.append("/0.jpg");
-
-		return sb.toString();
-	}
-
 	@Override
 	public String getPresetSize() {
 		if (_presetSize != null) {
@@ -133,17 +91,6 @@ public class BaseVideoEmbedderDisplayContext implements VideoEmbedderDisplayCont
 		return _presetSize;
 	}
 
-	public String getStartTime() {
-		if (_startTime != null) {
-			return _startTime;
-		}
-
-		_startTime = _portletPreferences.getValue(
-			"startTime", StringPool.BLANK);
-
-		return _startTime;
-	}
-
 	@Override
 	public String getURL() {
 		if (_url != null) {
@@ -153,11 +100,6 @@ public class BaseVideoEmbedderDisplayContext implements VideoEmbedderDisplayCont
 		_url = _portletPreferences.getValue("url", StringPool.BLANK);
 
 		return _url;
-	}
-
-	public String getWatchURL() {
-		return HttpUtil.getProtocol(_request) + "://www.youtube.com/watch?v=" +
-			getId();
 	}
 
 	@Override
@@ -180,39 +122,6 @@ public class BaseVideoEmbedderDisplayContext implements VideoEmbedderDisplayCont
 		return _width;
 	}
 
-	public boolean isAnnotations() {
-		if (_annotations != null) {
-			return _annotations;
-		}
-
-		_annotations = GetterUtil.getBoolean(
-			_portletPreferences.getValue("annotations", "true"));
-
-		return _annotations;
-	}
-
-	public boolean isAutoPlay() {
-		if (_autoPlay != null) {
-			return _autoPlay;
-		}
-
-		_autoPlay = GetterUtil.getBoolean(
-			_portletPreferences.getValue("autoplay", "false"));
-
-		return _autoPlay;
-	}
-
-	public boolean isClosedCaptioning() {
-		if (_closedCaptioning != null) {
-			return _closedCaptioning;
-		}
-
-		_closedCaptioning = GetterUtil.getBoolean(
-			_portletPreferences.getValue("closedCaptioning", "false"));
-
-		return _closedCaptioning;
-	}
-
 	@Override
 	public boolean isCustomSize() {
 		String presetSize = getPresetSize();
@@ -224,52 +133,12 @@ public class BaseVideoEmbedderDisplayContext implements VideoEmbedderDisplayCont
 		return false;
 	}
 
-	public boolean isEnableKeyboardControls() {
-		if (_enableKeyboardControls != null) {
-			return _enableKeyboardControls;
-		}
-
-		_enableKeyboardControls = GetterUtil.getBoolean(
-			_portletPreferences.getValue("enableKeyboardControls", "true"));
-
-		return _enableKeyboardControls;
-	}
-
-	public boolean isLoop() {
-		if (_loop != null) {
-			return _loop;
-		}
-
-		_loop = GetterUtil.getBoolean(
-			_portletPreferences.getValue("loop", "false"));
-
-		return _loop;
-	}
-
-	public boolean isShowThumbnail() {
-		if (_showThumbnail != null) {
-			return _showThumbnail;
-		}
-
-		_showThumbnail = GetterUtil.getBoolean(
-			_portletPreferences.getValue("showThumbnail", "false"));
-
-		return _showThumbnail;
-	}
-
-	private Boolean _annotations;
-	private Boolean _autoPlay;
-	private Boolean _closedCaptioning;
-	private Boolean _enableKeyboardControls;
-	private String _height;
-	private String _id;
-	private Boolean _loop;
-	private final PortletPreferences _portletPreferences;
-	private String _presetSize;
-	private final HttpServletRequest _request;
-	private Boolean _showThumbnail;
-	private String _startTime;
-	private String _url;
-	private String _width;
+	protected String _height;
+	protected String _id;
+	protected final PortletPreferences _portletPreferences;
+	protected String _presetSize;
+	protected final HttpServletRequest _request;
+	protected String _url;
+	protected String _width;
 
 }
